+++ backend/src/main/java/com/onair/hearit/application/HearitSearchService.java
@@ -1,7 +1,7 @@
 package com.onair.hearit.application;
 
 import com.onair.hearit.domain.Hearit;
-import com.onair.hearit.dto.request.TitleSearchRequest;
+import com.onair.hearit.dto.request.TitleSearchCondition;
 import com.onair.hearit.dto.response.HearitSearchResponse;
 import com.onair.hearit.infrastructure.HearitRepository;
 import java.util.List;
@@ -19,7 +19,7 @@ public class HearitSearchService {
         this.hearitRepository = hearitRepository;
     }
 
-    public List<HearitSearchResponse> searchByTitle(TitleSearchRequest condition) {
+    public List<HearitSearchResponse> searchByTitle(TitleSearchCondition condition) {
         Pageable pageable = PageRequest.of(condition.page(), condition.size());
         Page<Hearit> pageResult = hearitRepository.findByTitleOrderByCreatedAtDesc(condition.search(), pageable);
         return HearitSearchResponse.from(pageResult);
