+++ backend/src/main/java/com/onair/hearit/dto/request/TitleSearchCondition.java
@@ -5,8 +5,9 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public record TitleSearchRequest(
+public record TitleSearchCondition(
         @NotNull @Size(max = 20) String search,
         @NotNull @Min(0) Integer page,
         @NotNull @Min(0) @Max(20) Integer size
-) {}
+) {
+}
